Updated unit tests.
